MT47833: Move PlanningJobController contextmenu logic to PlanningJobTrait
